+++ engine/flink/generic/src/test/scala/pl/touk/nussknacker/genericmodel/RunGenericModelLocally.scala
@@ -1,8 +1,8 @@
 package pl.touk.nussknacker.genericmodel
 
 import com.typesafe.config.ConfigFactory
-import pl.touk.nussknacker.engine.ProcessManagerProvider
-import pl.touk.nussknacker.engine.testing.{LocalModelData, ProcessManagerProviderStub}
+import pl.touk.nussknacker.engine.DeploymentManagerProvider
+import pl.touk.nussknacker.engine.testing.{LocalModelData, DeploymentManagerProviderStub}
 import pl.touk.nussknacker.ui.util.LocalNussknackerWithSingleModel
 
 //Sample app to simplify local development.
@@ -13,7 +13,7 @@ object RunGenericModelLocally extends App {
 
   val managerConfig = ConfigFactory.empty()
   //For simplicity we use stub here, one can add real Flink implementation after add appropriate dependencies
-  val provider: ProcessManagerProvider = new ProcessManagerProviderStub
+  val provider: DeploymentManagerProvider = new DeploymentManagerProviderStub
   LocalNussknackerWithSingleModel.run(modelData, provider, managerConfig, Set("Default"))
 
 }