+++ engine/flink/management/src/it/scala/pl/touk/nussknacker/engine/management/streaming/StreamingDockerTest.scala
@@ -4,9 +4,9 @@ import com.typesafe.config.{Config, ConfigFactory, ConfigValueFactory}
 import com.typesafe.config.ConfigValueFactory.fromAnyRef
 import com.whisk.docker.{ContainerLink, DockerContainer, DockerReadyChecker}
 import org.scalatest.Suite
-import pl.touk.nussknacker.engine.api.deployment.ProcessManager
+import pl.touk.nussknacker.engine.api.deployment.DeploymentManager
 import pl.touk.nussknacker.engine.kafka.KafkaClient
-import pl.touk.nussknacker.engine.management.{DockerTest, FlinkStreamingProcessManagerProvider}
+import pl.touk.nussknacker.engine.management.{DockerTest, FlinkStreamingDeploymentManagerProvider}
 import pl.touk.nussknacker.engine.util.config.ScalaMajorVersionConfig
 
 import scala.concurrent.duration._
@@ -49,6 +49,6 @@ trait StreamingDockerTest extends DockerTest { self: Suite =>
 
   private def kafkaAddress = s"${ipOfContainer(kafkaContainer)}:$KafkaPort"
 
-  protected lazy val processManager: ProcessManager = FlinkStreamingProcessManagerProvider.defaultProcessManager(config)
+  protected lazy val deploymentManager: DeploymentManager = FlinkStreamingDeploymentManagerProvider.defaultDeploymentManager(config)
 
 }