+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/api/DefinitionResources.scala
@@ -48,7 +48,7 @@ class DefinitionResources(modelDataProvider: ProcessingTypeDataProvider[ModelDat
                 complete(
                   UIProcessObjectsFactory.prepareUIProcessObjects(
                     processingTypeData.modelData,
-                    processingTypeData.processManager,
+                    processingTypeData.deploymentManager,
                     user,
                     subprocesses,
                     isSubprocess,