+++ ui/server/src/test/scala/pl/touk/nussknacker/ui/definition/UIProcessObjectsFactorySpec.scala
@@ -12,7 +12,7 @@ import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
 import pl.touk.nussknacker.engine.{ModelData, ProcessingTypeConfig}
 import pl.touk.nussknacker.ui.api.helpers.TestFactory
-import pl.touk.nussknacker.ui.api.helpers.TestFactory.MockProcessManager
+import pl.touk.nussknacker.ui.api.helpers.TestFactory.MockDeploymentManager
 import pl.touk.nussknacker.ui.process.ConfigProcessCategoryService
 import pl.touk.nussknacker.ui.util.ConfigWithScalaVersion
 
@@ -62,7 +62,7 @@ class UIProcessObjectsFactorySpec extends FunSuite with Matchers {
 
   }
 
-  private val mockProcessManager = new MockProcessManager
+  private val mockDeploymentManager = new MockDeploymentManager
 
   test("should read editor from annotations") {
     val model: ModelData = LocalModelData(ConfigWithScalaVersion.streamingProcessTypeConfig, new EmptyProcessConfigCreator() {
@@ -72,7 +72,7 @@ class UIProcessObjectsFactorySpec extends FunSuite with Matchers {
 
     val processObjects = UIProcessObjectsFactory.prepareUIProcessObjects(
       model,
-      mockProcessManager,
+      mockDeploymentManager,
       TestFactory.user("userId"),
       Set(),
       false,
@@ -101,7 +101,7 @@ class UIProcessObjectsFactorySpec extends FunSuite with Matchers {
     })
 
     val processObjects =
-      UIProcessObjectsFactory.prepareUIProcessObjects(model, mockProcessManager, TestFactory.user("userId"), Set(), false,
+      UIProcessObjectsFactory.prepareUIProcessObjects(model, mockDeploymentManager, TestFactory.user("userId"), Set(), false,
         new ConfigProcessCategoryService(ConfigWithScalaVersion.config))
 
     processObjects.nodesToAdd.filter(_.name == "hiddenCategory") shouldBe empty
@@ -117,7 +117,7 @@ class UIProcessObjectsFactorySpec extends FunSuite with Matchers {
     })
 
     val processObjects =
-      UIProcessObjectsFactory.prepareUIProcessObjects(model, mockProcessManager, TestFactory.user("userId"), Set(), false,
+      UIProcessObjectsFactory.prepareUIProcessObjects(model, mockDeploymentManager, TestFactory.user("userId"), Set(), false,
         new ConfigProcessCategoryService(ConfigWithScalaVersion.config))
 
     val nodeGroups = processObjects.nodesToAdd.filter(_.name == "someCategory")